requirements.txt
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
APScheduler==3.10.4
google-api-python-client==2.97.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
gunicorn==21.2.0
azure-appconfiguration==1.5.0
azure-identity==1.15.0
Flask-Compress==1.14.0
pytz==2024.1 